fix(hierarchical-grid): fixed row island filtering
